requirements.txt
SpeechRecognition
pyttsx3
openai
python-dotenv
wikipedia
requests